Migrate conf client changes into existing client
